+++ src/lang/locale/ja.ts
@@ -7,6 +7,7 @@ export default {
     NEW_CARDS: "新規のカード",
     TOTAL_CARDS: "カード合計",
     BACK: "Back",
+    SKIP: "Skip",
     EDIT_CARD: "Edit Card",
     RESET_CARD_PROGRESS: "カードの進捗をリセット",
     HARD: "Hard",