+++ src/lang/locale/ko.ts
@@ -7,6 +7,7 @@ export default {
     NEW_CARDS: "새로운 카드들",
     TOTAL_CARDS: "전체 카드들",
     BACK: "Back",
+    SKIP: "Skip",
     EDIT_CARD: "Edit Card",
     RESET_CARD_PROGRESS: "카드의 진행상황을 초기화합니다.",
     HARD: "어려움(Hard)",